MSDN 의 제안대로 SendMessage 수정

+++ App/UpdaterForm.cs
@@ -8,7 +8,7 @@ namespace App
     public partial class UpdaterForm : Form
     {
         [DllImport("user32.dll")]
-        static extern int SendMessage(IntPtr hWnd, int Msg, int wParam, int lParam);
+        static extern IntPtr SendMessage(IntPtr hWnd, int Msg, IntPtr wParam, IntPtr lParam);
         [DllImport("user32.dll")]
         static extern bool ReleaseCapture();
 
@@ -25,7 +25,7 @@ namespace App
             if (e.Button == MouseButtons.Left)
             {
                 //ReleaseCapture();
-                //SendMessage(Handle, WM_NCLBUTTONDOWN, HT_CAPTION, 0);
+                //SendMessage(Handle, WM_NCLBUTTONDOWN, new IntPtr(HT_CAPTION), IntPtr.Zero);
             }
         }
